feat(linter): add rule `noDisabledTests`
