+++ crates/biome_js_formatter/src/js/expressions/template_expression.rs
@@ -4,7 +4,6 @@ use biome_formatter::write;
 use crate::js::expressions::static_member_expression::member_chain_callee_needs_parens;
 use crate::js::lists::template_element_list::FormatJsTemplateElementListOptions;
 use crate::parentheses::NeedsParentheses;
-use crate::utils::test_call::is_test_each_pattern;
 use biome_js_syntax::{AnyJsExpression, JsSyntaxNode, JsTemplateExpression, TsTemplateLiteralType};
 use biome_js_syntax::{JsSyntaxToken, TsTypeArguments};
 use biome_rowan::{declare_node_union, SyntaxResult};
@@ -69,7 +68,7 @@ impl AnyJsTemplate {
     fn write_elements(&self, f: &mut JsFormatter) -> FormatResult<()> {
         match self {
             AnyJsTemplate::JsTemplateExpression(template) => {
-                let is_test_each_pattern = is_test_each_pattern(template);
+                let is_test_each_pattern = template.is_test_each_pattern();
                 let options = FormatJsTemplateElementListOptions {
                     is_test_each_pattern,
                 };
