+++ crates/biome_js_formatter/src/utils/member_chain/mod.rs
@@ -114,7 +114,6 @@ use crate::utils::member_chain::groups::{
     MemberChainGroup, MemberChainGroupsBuilder, TailChainGroups,
 };
 pub use crate::utils::member_chain::simple_argument::SimpleArgument;
-use crate::utils::test_call::is_test_call_expression;
 use crate::JsLabels;
 use biome_formatter::{write, Buffer};
 use biome_js_syntax::{
@@ -363,7 +362,7 @@ impl Format<JsFormatContext> for MemberChain {
         if self.tail.len() <= 1 && !has_comments {
             return if is_long_curried_call(Some(&self.root)) {
                 write!(f, [format_one_line])
-            } else if is_test_call_expression(&self.root)? && self.head.members().len() >= 2 {
+            } else if self.root.is_test_call_expression()? && self.head.members().len() >= 2 {
                 write!(f, [self.head, soft_line_indent_or_space(&self.tail)])
             } else {
                 write!(f, [group(&format_one_line)])
