+++ crates/biome_js_formatter/src/utils/mod.rs
@@ -14,7 +14,6 @@ mod object_like;
 mod object_pattern_like;
 #[cfg(test)]
 mod quickcheck_utils;
-pub(crate) mod test_call;
 pub(crate) mod test_each_template;
 mod typescript;
