+++ website/src/content/docs/linter/rules/index.mdx
@@ -232,7 +232,7 @@ Nursery rules get promoted to other groups once they become stable or may be rem
 Rules that belong to this group <strong>are not subject to semantic version</strong>.
 | Rule name | Description | Properties |
 | --- | --- | --- |
-| [noDisabledTests](/linter/rules/no-disabled-tests) | Succinct description of the rule. |  |
+| [noDisabledTests](/linter/rules/no-disabled-tests) | Disallow disabled tests. |  |
 | [noDuplicateJsonKeys](/linter/rules/no-duplicate-json-keys) | Disallow two keys with the same name inside a JSON object. |  |
 | [noEmptyBlockStatements](/linter/rules/no-empty-block-statements) | Disallow empty block statements and static blocks. |  |
 | [noEmptyTypeParameters](/linter/rules/no-empty-type-parameters) | Disallow empty type parameters in type aliases and interfaces. |  |
