+++ xtask/codegen/src/generate_new_lintrule.rs
@@ -36,8 +36,6 @@ declare_rule! {{
     ///
     /// Add a link to the corresponding ESLint rule (if any):
     ///
-    /// Source: https://eslint.org/docs/latest/rules/rule-name
-    ///
     /// ## Examples
     ///
     /// ### Invalid
